Add separate command for disassembling modules.
