+++ src/commandline/commandline.h
@@ -18,6 +18,15 @@
 
 #include "package.h"
 
+/*
+ * Forward declarations.
+ */
+
+namespace slang::interpreter
+{
+class context;
+}    // namespace slang::interpreter
+
 namespace slang::commandline
 {
 
@@ -162,5 +171,26 @@ public:
     void invoke(const std::vector<std::string>& args) override;
     std::string get_description() const override;
 };
+
+/** Bytecode disassembly. */
+class disasm : public command
+{
+    /** The package manager bound to this command. */
+    slang::package_manager& manager;
+
+public:
+    /** Constructor. */
+    explicit disasm(slang::package_manager& manager);
+    void invoke(const std::vector<std::string>& args) override;
+    std::string get_description() const override;
+};
+
+/**
+ * Set up the default runtime environment for a context.
+ *
+ * @param ctx The context to set up.
+ * @param verbose Whether to enable verbose logging. Defaults to `false`.
+ */
+void runtime_setup(slang::interpreter::context& ctx, bool verbose = false);
 
 }    // namespace slang::commandline
